refactor: --save-peer
