+++ packages/cli-utils/src/getSaveType.ts
@@ -1,8 +1,10 @@
 import { Config } from '@pnpm/config'
 import { DependenciesField } from '@pnpm/types'
 
-export default function getSaveType (opts: Pick<Config, 'saveDev' | 'saveOptional' | 'saveProd'>): DependenciesField | undefined {
-  if (opts.saveDev) return 'devDependencies'
+export default function getSaveType (
+  opts: Pick<Config, 'saveDev' | 'saveOptional' | 'saveProd' | 'savePeer'>,
+): DependenciesField | undefined {
+  if (opts.saveDev || opts.savePeer) return 'devDependencies'
   if (opts.saveOptional) return 'optionalDependencies'
   if (opts.saveProd) return 'dependencies'
   return undefined
