+++ packages/plugin-commands-installation/package.json
@@ -44,6 +44,7 @@
     "@types/ramda": "^0.26.40",
     "@types/sinon": "^7.5.1",
     "@types/table": "^4.0.7",
+    "load-json-file": "6.2.0",
     "make-dir": "3.0.0",
     "proxyquire": "2.1.3",
     "read-yaml-file": "1.1.0",
